refactoring after first comments
